Reorganize module comments and some POD
Move the single copyright notice to the end of the POD and add the
magic Emacs variables so that Emacs will find it there for automatic
updates.  Add an SPDX license identifier and remove the duplicate
copy of the license in the leading comment.  Update the style of
the leading comment for my current Perl style.  Reorder some of the
POD sections to match my current style.