Add a new object-oriented API
Add a new object-oriented API.  This avoids global variables and
global state, provides a simpler API that doesn't require or support
the now-obsolete version string, and calls croak on errors.  The
legacy API of pgp_sign, pgp_verify, and pgp_error is still supported
but is now implemented in terms of the object-oriented API.  The new
API should be used for all new code.

pgp_sign, when called in array context, now always returns "GnuPG" as
the version string, and the version passed into pgp_verify is always
ignored.  Including the OpenPGP implementation version information in
signatures is obsolete; GnuPG no longer does it by default and it
serves no useful purpose.

Rewrite most of the tests to use the new API.  Add checks that GnuPG
binaries are available and skip tests if they aren't.