Bump test requirements to GnuPG 2.1.23 and 1.4.20
Document that GnuPG 2.1.23 or GnuPG 1.4.20 or later is required and
skip tests on platforms that do not meet those version requirements.
The alternative would be auto-discovery of which command-line flags to
use and these version requirements are met by Debian stable (and
Debian oldstable with backports), so hopefully this restriction will
not cause too much hardship.