Fix signing with large data
PGP::Sign 1.00 and later failed to sign input that's larger than can
be held by the operating system pipe buffer because IPC::Run set the
input pipe to non-blocking.  Fix this by using a non-blocking write in
a select loop, and avoid a deadlock by using a pipe for the passphrase
as well instead of a scalar.